refactor : extract function to lib file
